Restructure services files
