+++ commands/index.js
@@ -1,9 +1,8 @@
+//TODO: Add a loop to import all of these instead of manually importing each one
 exports.getApplicationCommands = () => {
   return {
-    //Timer
     goats: require('./goats/'),
     remind: require('./remind'),
-    //Hub
     help: require('./help'),
     about: require('./about'),
     sheets: require('./sheets'),
